+++ src/main/java/net/geforcemods/securitycraft/blockentities/CageTrapBlockEntity.java
@@ -31,6 +31,7 @@ public class CageTrapBlockEntity extends DisguisableBlockEntity {
 		return disabled.get();
 	}
 
+	@Override
 	public boolean ignoresOwner() {
 		return ignoreOwner.get();
 	}