+++ src/main/java/net/geforcemods/securitycraft/blockentities/RiftStabilizerBlockEntity.java
@@ -325,6 +325,7 @@ public class RiftStabilizerBlockEntity extends DisguisableBlockEntity implements
 		}
 	}
 
+	@Override
 	public boolean ignoresOwner() {
 		return ignoreOwner.get();
 	}