+++ src/main/java/net/geforcemods/securitycraft/blockentities/TrophySystemBlockEntity.java
@@ -342,6 +342,7 @@ public class TrophySystemBlockEntity extends DisguisableBlockEntity implements I
 		return disabled.get();
 	}
 
+	@Override
 	public boolean ignoresOwner() {
 		return ignoreOwner.get();
 	}