+++ src/main/java/net/geforcemods/securitycraft/blockentities/UsernameLoggerBlockEntity.java
@@ -125,6 +125,7 @@ public class UsernameLoggerBlockEntity extends DisguisableBlockEntity implements
 		return disabled.get();
 	}
 
+	@Override
 	public boolean ignoresOwner() {
 		return ignoreOwner.get();
 	}