+++ src/main/java/net/geforcemods/securitycraft/misc/TargetingMode.java
@@ -1,5 +1,10 @@
 package net.geforcemods.securitycraft.misc;
 
+import net.geforcemods.securitycraft.api.IModuleInventory;
+import net.geforcemods.securitycraft.api.IOwnable;
+import net.geforcemods.securitycraft.util.EntityUtils;
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.util.text.ITextComponent;
 import net.minecraft.util.text.TranslationTextComponent;
 
@@ -24,5 +29,19 @@ public enum TargetingMode {
 
 	public boolean allowsMobs() {
 		return this == MOBS || this == PLAYERS_AND_MOBS;
+	}
+
+	public <T extends IOwnable> boolean canAttackEntity(LivingEntity entity, T be, boolean checkInvisibility) {
+		if (entity == null || be instanceof IModuleInventory && ((IModuleInventory) be).isAllowed(entity))
+			return false;
+
+		boolean isPlayer = entity instanceof PlayerEntity;
+
+		if (isPlayer && allowsPlayers() || !isPlayer && allowsMobs()) {
+			return (!isPlayer || !(be.isOwnedBy((PlayerEntity) entity) && be.ignoresOwner()) && !((PlayerEntity) entity).isCreative()) //Player checks
+					&& !entity.isSpectator() && (!checkInvisibility || !EntityUtils.isInvisible(entity)) && !be.allowsOwnableEntity(entity); //checks for all entities
+		}
+		else
+			return false;
 	}
 }